Make sure MPIR_Breakpoint() is compiled without CFLAGS.
In optimized builds, CFLAGS contains various optimizations such as -O3,
and is propogated by automake to all files. To work-around this,
isolate MPIR_Breakpoint() and other MPIR_* symbols into its own library
built with debugger specific CFLAGS.

To prevent CFLAGS from being polluted elsewhere in the make tree, build
this in its own tiny stand-alone makefile.

Fixes #7757

(cherry picked from commit 6d820037e92988197e10264a810ebdb4f2b8aebc)
