+++ orte/orted/Makefile.am
@@ -12,7 +12,7 @@
 #                         All rights reserved.
 # Copyright (c) 2014      Cisco Systems, Inc.  All rights reserved.
 # Copyright (c) 2015      Intel, Inc. All rights reserved.
-# Copyright (c) 2018      IBM Corporation.  All rights reserved.
+# Copyright (c) 2018-2021 IBM Corporation.  All rights reserved.
 # $COPYRIGHT$
 #
 # Additional copyrights may follow
@@ -30,17 +30,7 @@ headers += \
 
 lib@ORTE_LIB_PREFIX@open_rte_la_SOURCES += \
         orted/orted_main.c \
-        orted/orted_comm.c
-
-# The MPIR portion of the library must be built with -g, even if
-# the rest of the library has other optimization flags.
-# Use an intermediate library to isolate the debug object.
-noinst_LTLIBRARIES += liborted_mpir.la
-liborted_mpir_la_SOURCES = \
+        orted/orted_comm.c \
 	orted/orted_submit.c
-liborted_mpir_la_CFLAGS = $(CFLAGS_WITHOUT_OPTFLAGS) $(DEBUGGER_CFLAGS)
-
-lib@ORTE_LIB_PREFIX@open_rte_la_LIBADD += liborted_mpir.la
-
 
 include orted/pmix/Makefile.am
